// abrpoint.client/src/models/PointeuseModel.ts
export type Pointeuse = {
    poicod: string;
    soccod: string;
    poilib?: string;
    poiadrip1?: number;
    poiadrip2?: number;
    poiadrip3?: number;
    poiadrip4?: number;
    poiport?: number;
    poietat?: string;
    poicom?: string;
  }
  export type PointeusesCred = {
    ips:string[],
    //pswd:string
  }
  